third_party/WebKit/LayoutTests/platform/win/external/wpt/workers/opaque-origin-expected.txt

This is a testharness.js-based test.
PASS Worker has an opaque origin. 
FAIL Worker can read its own blobs. assert_equals: expected (string) "from worker" but got (object) null
FAIL Worker can read its owners blobs. assert_equals: expected (string) "from page" but got (object) null
FAIL Worker can XHR fetch a blob. assert_equals: expected "from page" but got ""
FAIL Worker can fetch a blob. promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
FAIL Worker can access BroadcastChannel Failed to construct 'BroadcastChannel': Can't create BroadcastChannel in an opaque origin
Harness: the test ran to completion.
